LPS-180956 Extract COREntryDetailsScreenNavigationCategory methods to a COREntryDetailsScreenNavigationEntry component and extend COREntryDetailsScreenNavigationCategory in it
