+++ modules/apps/commerce/commerce-order-rule-web/src/main/java/com/liferay/commerce/order/rule/web/internal/frontend/taglib/servlet/taglib/COREntryDetailsScreenNavigationCategory.java
@@ -14,34 +14,13 @@
 
 package com.liferay.commerce.order.rule.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.currency.service.CommerceCurrencyService;
-import com.liferay.commerce.order.rule.entry.type.COREntryTypeJSPContributorRegistry;
-import com.liferay.commerce.order.rule.entry.type.COREntryTypeRegistry;
-import com.liferay.commerce.order.rule.model.COREntry;
-import com.liferay.commerce.order.rule.service.COREntryService;
-import com.liferay.commerce.order.rule.web.internal.display.context.COREntryDisplayContext;
 import com.liferay.commerce.order.rule.web.internal.entry.constants.COREntryScreenNavigationEntryConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -50,22 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class COREntryDetailsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<COREntry> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return COREntryScreenNavigationEntryConstants.CATEGORY_KEY_DETAILS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return COREntryScreenNavigationEntryConstants.CATEGORY_KEY_DETAILS;
 	}
 
@@ -74,7 +45,7 @@ public class COREntryDetailsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(
+		return language.get(
 			resourceBundle,
 			COREntryScreenNavigationEntryConstants.CATEGORY_KEY_DETAILS);
 	}
@@ -85,75 +56,7 @@ public class COREntryDetailsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COR_ENTRY_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, COREntry corEntry) {
-		if (corEntry == null) {
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission = _corEntryModelResourcePermission.contains(
-				PermissionThreadLocal.getPermissionChecker(),
-				corEntry.getCOREntryId(), ActionKeys.UPDATE);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		COREntryDisplayContext corEntryDisplayContext =
-			new COREntryDisplayContext(
-				_commerceCurrencyService, _corEntryModelResourcePermission,
-				_corEntryService, _corEntryTypeJSPContributorRegistry,
-				_corEntryTypeRegistry, httpServletRequest, _portal);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, corEntryDisplayContext);
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse, "/cor_entry/details.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		COREntryDetailsScreenNavigationCategory.class);
-
 	@Reference
-	private CommerceCurrencyService _commerceCurrencyService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.order.rule.model.COREntry)"
-	)
-	private ModelResourcePermission<COREntry> _corEntryModelResourcePermission;
-
-	@Reference
-	private COREntryService _corEntryService;
-
-	@Reference
-	private COREntryTypeJSPContributorRegistry
-		_corEntryTypeJSPContributorRegistry;
-
-	@Reference
-	private COREntryTypeRegistry _corEntryTypeRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private Portal _portal;
+	protected Language language;
 
 }
